requirement.txt
langchain_google_genai
streamlit
python-dotenv
pyPDF2
langchain


-e .